[EQLMatch] restructuring, created match.py, and quantify_entity.py.
