+++ test/test_eql/test_aggregations.py
@@ -1,14 +1,13 @@
 from krrood.entity_query_language.entity import (
     flatten,
     entity,
-    an,
     not_,
     in_,
-    the,
     for_all,
     let,
     exists,
 )
+from krrood.entity_query_language.quantify_entity import an, the
 from ..dataset.example_classes import VectorsWithProperty
 from ..dataset.semantic_world_like_classes import View, Drawer, Container, Cabinet
