+++ test/test_eql/test_core/test_rules.py
@@ -1,5 +1,6 @@
 from krrood.entity_query_language.conclusion import Add
-from krrood.entity_query_language.entity import let, an, entity, and_, inference
+from krrood.entity_query_language.entity import let, entity, and_, inference
+from krrood.entity_query_language.quantify_entity import an
 from krrood.entity_query_language.predicate import HasType
 from krrood.entity_query_language.rule import refinement, alternative, next_rule
 from ...dataset.semantic_world_like_classes import (
