+++ test/test_eql/test_symbol_graph.py
@@ -2,7 +2,8 @@ import os
 
 import pytest
 
-from krrood.entity_query_language.entity import an, entity, let
+from krrood.entity_query_language.entity import entity, let
+from krrood.entity_query_language.quantify_entity import an
 from krrood.entity_query_language.symbol_graph import SymbolGraph
 from ..dataset.example_classes import Position
